requirements.txt
astroid==1.6.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==2.0
django-extensions==1.9.8
django-filter==1.1.0
django-heroku==0.2.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.10
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.3.2
pylint==1.8.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
yapf==0.20.0
firebase-admin==2.7.0
wolframalpha==3.0
django-cloudinary-storage==0.2.3
Pillow==5.0.0
django-froala-editor==2.7.5
pyfcm==1.4.4
arrow==0.12.1
